Very simple compatibility implementation
API should now be able to work for Framework-version 3.0.0 - 3.0.3
