+++ src/main/java/org/utplsql/api/testRunner/ActualTestRunnerStatement.java
@@ -1,0 +1,39 @@
+package org.utplsql.api.testRunner;
+
+import org.utplsql.api.TestRunnerOptions;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+/** Provides the call to run tests for the most actual Framework version.
+ * Includes fail on error
+ *
+ * @author  pesse
+ */
+public class ActualTestRunnerStatement extends AbstractTestRunnerStatement {
+
+    public ActualTestRunnerStatement(TestRunnerOptions options, Connection connection ) throws SQLException {
+        super( options, connection);
+    }
+
+    @Override
+    protected String getSql() {
+        // Workaround because Oracle JDBC doesn't support passing boolean to stored procedures.
+        String colorConsoleStr = Boolean.toString(options.colorConsole);
+        String failOnErrors = Boolean.toString(options.failOnErrors);
+
+        return
+                "BEGIN " +
+                        "ut_runner.run(" +
+                        "a_paths                => ?, " +
+                        "a_reporters            => ?, " +
+                        "a_color_console        => " + colorConsoleStr + ", " +
+                        "a_coverage_schemes     => ?, " +
+                        "a_source_file_mappings => ?, " +
+                        "a_test_file_mappings   => ?, " +
+                        "a_include_objects      => ?, " +
+                        "a_exclude_objects      => ?, " +
+                        "a_fail_on_errors       => " + failOnErrors + "); " +
+                        "END;";
+    }
+}
