+++ src/main/java/org/utplsql/api/testRunner/TestRunnerStatement.java
@@ -1,34 +1,14 @@
 package org.utplsql.api.testRunner;
 
-import org.utplsql.api.TestRunnerOptions;
-
-import java.sql.Connection;
 import java.sql.SQLException;
 
-public class TestRunnerStatement extends AbstractTestRunnerStatement {
+/** Interface to hide the concrete Statement-implementations of TestRunner
+ *
+ * @author pesse
+ */
+public interface TestRunnerStatement {
 
-    public TestRunnerStatement(TestRunnerOptions options, Connection connection ) throws SQLException {
-        super( options, connection);
-    }
+    void execute() throws SQLException;
 
-    @Override
-    protected String getSql() {
-        // Workaround because Oracle JDBC doesn't support passing boolean to stored procedures.
-        String colorConsoleStr = Boolean.toString(options.colorConsole);
-        String failOnErrors = Boolean.toString(options.failOnErrors);
-
-        return
-                "BEGIN " +
-                        "ut_runner.run(" +
-                        "a_paths                => ?, " +
-                        "a_reporters            => ?, " +
-                        "a_color_console        => " + colorConsoleStr + ", " +
-                        "a_coverage_schemes     => ?, " +
-                        "a_source_file_mappings => ?, " +
-                        "a_test_file_mappings   => ?, " +
-                        "a_include_objects      => ?, " +
-                        "a_exclude_objects      => ?, " +
-                        "a_fail_on_errors       => " + failOnErrors + "); " +
-                        "END;";
-    }
+    void close() throws SQLException;
 }
